# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ORBITOS API",
    "version": "0.1.0"
  },
  "paths": {
    "/orbitos-api/v1/electrometer/{device_id}/connect/{ip}": {
      "post": {
        "tags": [
          "electrometer"
        ],
        "summary": "Connect To Electrometer",
        "description": "Connect to the electrometer with the given device ID.",
        "operationId": "electrometer-connect_to_electrometer",
        "parameters": [
          {
            "name": "ip",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ip"
            }
          },
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/orbitos-api/v1/electrometer/{device_id}/disconnect": {
      "post": {
        "tags": [
          "electrometer"
        ],
        "summary": "Disconnect Electrometer",
        "description": "Disconnect the electrometer.",
        "operationId": "electrometer-disconnect_electrometer",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/orbitos-api/v1/electrometer/electrometers/reset": {
      "post": {
        "tags": [
          "electrometer"
        ],
        "summary": "Reset Electrometer",
        "description": "Reset all electrometers.",
        "operationId": "electrometer-reset_electrometer",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/orbitos-api/v1/electrometer/{device_id}/state": {
      "get": {
        "tags": [
          "electrometer"
        ],
        "summary": "Get Electrometer State",
        "description": "Get the current state of the electrometer.",
        "operationId": "electrometer-get_electrometer_state",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ElectrometerState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/orbitos-api/v1/electrometer/{device_id}/settings": {
      "get": {
        "tags": [
          "electrometer"
        ],
        "summary": "Get Electrometer Settings",
        "description": "Get the current settings of the electrometer.",
        "operationId": "electrometer-get_electrometer_settings",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ElectrometerSettings"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "electrometer"
        ],
        "summary": "Set Electrometer Settings",
        "description": "Set the one or more setting of the electrometer.",
        "operationId": "electrometer-set_electrometer_settings",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ElectrometerSettingsSet"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/orbitos-api/v1/electrometer/{device_id}/state/reset-error": {
      "post": {
        "tags": [
          "electrometer"
        ],
        "summary": "Reset Electrometer Error",
        "description": "Reset the error state of the electrometer.",
        "operationId": "electrometer-reset_electrometer_error",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/orbitos-api/v1/electrometer/{device_id}/data": {
      "get": {
        "tags": [
          "electrometer"
        ],
        "summary": "Get Current Data",
        "description": "Get the current data from the electrometer for a specified time frame.",
        "operationId": "electrometer-get_current_data",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ElectrometerDataResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/orbitos-api/v1/electrometer/{device_id}/continuous-measurement/start": {
      "post": {
        "tags": [
          "electrometer"
        ],
        "summary": "Start Continuous Measurement",
        "description": "Start continuous measurement on the electrometer.",
        "operationId": "electrometer-start_continuous_measurement",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/orbitos-api/v1/electrometer/{device_id}/continuous-measurement/stop": {
      "post": {
        "tags": [
          "electrometer"
        ],
        "summary": "Stop Continuous Measurement",
        "description": "Stop continuous measurement on the electrometer.",
        "operationId": "electrometer-stop_continuous_measurement",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/orbitos-api/v1/electrometer/{device_id}/trigger-based-measurement/initialize": {
      "post": {
        "tags": [
          "electrometer"
        ],
        "summary": "Initialize Trigger Based Measurement",
        "description": "Initialize trigger-based measurement on the electrometer.",
        "operationId": "electrometer-initialize_trigger_based_measurement",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/orbitos-api/v1/electrometer/{device_id}/trigger-based-measurement/start": {
      "post": {
        "tags": [
          "electrometer"
        ],
        "summary": "Start Trigger Based Measurement",
        "description": "Start trigger-based measurement on the electrometer.",
        "operationId": "electrometer-start_trigger_based_measurement",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/orbitos-api/v1/types/websocket": {
      "get": {
        "tags": [
          "types"
        ],
        "summary": "Get Websocket Type",
        "operationId": "types-get_websocket_type",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseWebSocketMessage"
                }
              }
            }
          }
        }
      }
    },
    "/orbitos-api/v1/types/base-state": {
      "get": {
        "tags": [
          "types"
        ],
        "summary": "Get Base State Type",
        "operationId": "types-get_base_state_type",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseState"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BaseResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "BaseResponse"
      },
      "BaseState": {
        "properties": {
          "device_name": {
            "type": "string",
            "title": "Device Name"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "unknown"
          },
          "connection_status": {
            "$ref": "#/components/schemas/ConnectionStatus",
            "default": "disconnected"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "device_name"
        ],
        "title": "BaseState",
        "description": "Base class for all state models."
      },
      "BaseWebSocketMessage": {
        "properties": {
          "type": {
            "$ref": "#/components/schemas/WebSocketMessageType"
          },
          "device_name": {
            "type": "string",
            "title": "Device Name"
          },
          "content": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "type",
          "device_name"
        ],
        "title": "BaseWebSocketMessage"
      },
      "ConnectionStatus": {
        "type": "string",
        "enum": [
          "connected",
          "connecting",
          "disconnected",
          "health check failed"
        ],
        "title": "ConnectionStatus"
      },
      "ElectrometerDataResponse": {
        "properties": {
          "device_name": {
            "$ref": "#/components/schemas/ElectrometerName"
          },
          "current": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Current"
          },
          "timestamp": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "device_name",
          "current",
          "timestamp"
        ],
        "title": "ElectrometerDataResponse"
      },
      "ElectrometerName": {
        "type": "string",
        "enum": [
          "electrometer_1",
          "electrometer_2"
        ],
        "title": "ElectrometerName"
      },
      "ElectrometerSettings": {
        "properties": {
          "device_id": {
            "type": "integer",
            "title": "Device Id"
          },
          "trigger_count": {
            "type": "integer",
            "title": "Trigger Count",
            "default": 100
          },
          "trigger_time_interval": {
            "type": "number",
            "title": "Trigger Time Interval",
            "default": 0.1
          },
          "trigger_bypass": {
            "type": "string",
            "title": "Trigger Bypass",
            "default": "OFF"
          },
          "trigger_delay": {
            "type": "number",
            "title": "Trigger Delay",
            "default": 0.0
          },
          "function": {
            "type": "string",
            "title": "Function",
            "default": "CURR"
          },
          "aperture_integration_time": {
            "type": "number",
            "title": "Aperture Integration Time",
            "default": 0.2
          },
          "aperture_auto": {
            "type": "string",
            "title": "Aperture Auto",
            "default": "OFF"
          },
          "current_range": {
            "type": "number",
            "title": "Current Range",
            "default": 1e-07
          },
          "current_range_auto": {
            "type": "string",
            "title": "Current Range Auto",
            "default": "ON"
          },
          "current_range_auto_upper_limit": {
            "type": "number",
            "title": "Current Range Auto Upper Limit",
            "default": 1e-07
          },
          "current_range_auto_lower_limit": {
            "type": "number",
            "title": "Current Range Auto Lower Limit",
            "default": 1e-16
          }
        },
        "type": "object",
        "required": [
          "device_id"
        ],
        "title": "ElectrometerSettings"
      },
      "ElectrometerSettingsSet": {
        "properties": {
          "trigger_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trigger Count"
          },
          "trigger_time_interval": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trigger Time Interval"
          },
          "trigger_bypass": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trigger Bypass"
          },
          "trigger_delay": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trigger Delay"
          },
          "function": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Function"
          },
          "aperture_integration_time": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aperture Integration Time"
          },
          "aperture_auto": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aperture Auto"
          },
          "current_range": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Range"
          },
          "current_range_auto": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Range Auto"
          },
          "current_range_auto_upper_limit": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Range Auto Upper Limit"
          },
          "current_range_auto_lower_limit": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Range Auto Lower Limit"
          }
        },
        "type": "object",
        "title": "ElectrometerSettingsSet"
      },
      "ElectrometerState": {
        "properties": {
          "device_name": {
            "type": "string",
            "title": "Device Name"
          },
          "status": {
            "$ref": "#/components/schemas/ElectrometerStatus",
            "default": "unknown"
          },
          "connection_status": {
            "$ref": "#/components/schemas/ConnectionStatus",
            "default": "disconnected"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "device_name"
        ],
        "title": "ElectrometerState"
      },
      "ElectrometerStatus": {
        "type": "string",
        "enum": [
          "unknown",
          "idle",
          "trigger_based_measurement_running",
          "fetching_trigger_based_measurement_data",
          "waiting_to_start_continuous_measurement",
          "continuous_measurement_running"
        ],
        "title": "ElectrometerStatus"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WebSocketMessageType": {
        "type": "string",
        "enum": [
          "state",
          "data",
          "settings"
        ],
        "title": "WebSocketMessageType"
      }
    }
  }
}